requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
gunicorn==20.1.0
pycodestyle==2.8.0
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==6.0.0
psycopg2-binary==2.9.3

decouple~=0.0.7